Implement rebalacing in library

+++ src/Portfolio/Portfolio.fs
@@ -1,6 +1,6 @@
 module Portfolio
 
-type FundAmount(name: string, amount: int) =
+type FundAmount(name: string, amount: float) =
     member this.Amount = amount
     member this.Name = name
 
@@ -8,6 +8,18 @@ type Portfolio(equityFunds: FundAmount list, distribution: Map<string, float>) =
     member this.EquityFunds = equityFunds
     member this.Distribution = distribution
 
-
 let sumOfAssets funds =
     funds |> List.sumBy(fun (x:FundAmount) -> x.Amount)
+
+
+
+
+
+
+   
+
+
+
+
+
+
